email: send mail on registration
